detection_model_test/requirements.txt
accelerate==1.10.1
altair==5.5.0
annotated-types==0.7.0
anywidget==0.9.18
asttokens==3.0.0
attrs==25.4.0
bbox_visualizer==0.2.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.4
decorator==5.2.1
defusedxml==0.7.1
einops==0.8.1
executing==2.2.1
fairscale==0.4.13
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.7
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_bbox_widget==0.6.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==2.7.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
open_clip_torch==3.2.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pi_heif==1.1.1
pillow==11.3.0
pillow-avif-plugin==1.5.2
pip==24.3.1
polars==1.34.0
polars-runtime-32==1.34.0
polygraphy==0.49.26
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
psygnal==0.14.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycocotools==2.0.10
pydantic==2.12.0
pydantic_core==2.41.1
pydeck==0.9.1
Pygments==2.19.2
pylabel==0.1.55
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rf100vl==1.1.0
rfdetr==1.3.0
roboflow==1.2.10
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.50.0
supervision==0.26.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
timm==1.0.20
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ultralytics==8.3.211
ultralytics-thop==2.0.17
urllib3==2.5.0
wcwidth==0.2.14
widgetsnbextension==4.0.14
